Basic functionality... Gotta stop branching for small changes like this. Oh well, it's good practice for when we get to a good point actually
